automation/requirements.txt
anthropic==0.25.0
requests==2.31.0
python-dotenv==1.0.0
pytz==2024.1
